+++ src/blueetl/adapters/bluepy/__init__.py
@@ -1,4 +1,5 @@
 """Bluepy implementation."""
+
 from blueetl.utils import import_optional_dependency
 
 # Immediately raise an error with a custom message if the submodule is imported