+++ src/blueetl/extract/report.py
@@ -1,4 +1,5 @@
 """Generic Report extractor."""
+
 import logging
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass