+++ src/blueetl/extract/simulations.py
@@ -1,4 +1,5 @@
 """Simulations extractor."""
+
 import logging
 from enum import Enum
 from typing import Any, Optional, cast