+++ src/blueetl/types.py
@@ -1,4 +1,5 @@
 """Common types."""
+
 from os import PathLike
 from typing import Union